Abstract the query that we make to get announcement dates.
I'm still coming to term with using unique keys for the date data, but this feels like it will be a faster query then trying to query meta_value all the time. I could be wrong and may come back and change this at some point, but for now it works.

We'll use this in the post type list table columns as well.