Round about queries to find only dates with valid announcements.
First query for any posts with this month's meta key
Then query for posts that are published or private from the results of the previous query.
Finally, query for distinct meta keys for the published posts so that we can display what we want.

We should remove private from the query as we don't have much use for it.